+++ src/main/java/ticketaka/mtvs3_final_backend/redis/refreshtoken/repository/RefreshTokenRedisRepository.java
@@ -1,7 +1,7 @@
-package ticketaka.mtvs3_final_backend.redis.repository;
+package ticketaka.mtvs3_final_backend.redis.refreshtoken.repository;
 
 import org.springframework.data.repository.CrudRepository;
-import ticketaka.mtvs3_final_backend.redis.domain.RefreshToken;
+import ticketaka.mtvs3_final_backend.redis.refreshtoken.domain.RefreshToken;
 
 public interface RefreshTokenRedisRepository extends CrudRepository<RefreshToken, String> {
